event_management_backend/requirements.txt
apispec==6.8.1
blinker==1.9.0
click==8.1.8
flake8==7.2.0
Flask==3.1.0
flask-smorest==0.45.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
pycodestyle==2.13.0
pyflakes==3.3.2
pytest==8.3.5
webargs==8.6.0
Werkzeug==3.1.3
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.10
flask-jwt-extended==4.7.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
